--- rir_dataset_generator/requirements.txt
@@ -1,1 +1,0 @@
-rir-generator
