+++ scripts/test_data_set.py
@@ -1,8 +1,8 @@
 import torch
-import Utilities
+from scripts import Utilities
 import os
 
-from rir_dataset import RIR_DATASET
+from acustic_locating_vq_vae.rir_dataset_generator.rir_dataset import RIR_DATASET
 
 git_root_path = Utilities.get_git_root()
 dataset = RIR_DATASET(root_dir=os.path.join(git_root_path, 'rir_dataset_generator', 'dev_data'))
