+++ src/acustic_locating_vq_vae/vq_vae/convolutional_vq_vae.py
