+++ src/acustic_locating_vq_vae/vq_vae/modules/residual.py
