+++ src/acustic_locating_vq_vae/vq_vae/modules/residual_stack.py
@@ -25,7 +25,7 @@
  #   SOFTWARE.                                                                       #
  #####################################################################################
 
-from modules.residual import Residual
+from src.acustic_locating_vq_vae.vq_vae.modules.residual import Residual
 
 import torch.nn as nn
 import torch.nn.functional as F
