+++ src/acustic_locating_vq_vae/vq_vae/vector_quantizer.py
